Move request status JSPs and actions to specimen-main
