+++ api/src/org/labkey/api/study/SpecimenUrls.java
@@ -17,6 +17,7 @@
 package org.labkey.api.study;
 
 import org.labkey.api.action.UrlProvider;
+import org.labkey.api.annotations.Migrate;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.view.ActionURL;
@@ -30,10 +31,10 @@ public interface SpecimenUrls extends UrlProvider
     ActionURL getCommentURL(Container c, String globalUniqueId);
     ActionURL getDeleteRequestURL(Container c, String id);
     ActionURL getInsertSpecimenQueryRowURL(Container c, String schemaName, TableInfo table);
-    ActionURL getManageRequestURL(Container c);
-    ActionURL getManageRequestStatusURL(Container c, int requestId);
-    ActionURL getRequestDetailsURL(Container c, String requestId);
+    ActionURL getManageRequestURL(Container c, int requestId);
+    @Migrate // Eliminate this -- same as getManageRequestURL(Container, int);
     ActionURL getRequestDetailsURL(Container c, int requestId);
+    ActionURL getManageRequestStatusURL(Container c, int requestId);
     ActionURL getSpecimensURL(Container c);
     ActionURL getSpecimensURL(Container c, boolean showVials);
     ActionURL getSelectedSpecimensURL(Container c, boolean showVials);
