+++ study/src/org/labkey/study/controllers/specimen/SpecimenController.java
@@ -20,42 +20,30 @@ import org.apache.commons.io.Charsets;
 import org.apache.commons.lang3.time.DateUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.jetbrains.annotations.Nullable;
 import org.labkey.api.action.FormHandlerAction;
 import org.labkey.api.action.FormViewAction;
 import org.labkey.api.action.QueryViewAction;
 import org.labkey.api.action.SimpleRedirectAction;
 import org.labkey.api.action.SimpleViewAction;
-import org.labkey.api.action.SpringActionController;
 import org.labkey.api.attachments.AttachmentFile;
 import org.labkey.api.attachments.ByteArrayAttachmentFile;
-import org.labkey.api.audit.TransactionAuditProvider;
-import org.labkey.api.data.ActionButton;
 import org.labkey.api.data.BaseColumnInfo;
 import org.labkey.api.data.ButtonBar;
 import org.labkey.api.data.ColumnInfo;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.DataRegionSelection;
 import org.labkey.api.data.ExcelWriter;
-import org.labkey.api.data.MenuButton;
 import org.labkey.api.data.SimpleFilter;
 import org.labkey.api.data.TSVGridWriter;
 import org.labkey.api.data.TSVWriter;
-import org.labkey.api.data.Table;
 import org.labkey.api.data.TableInfo;
-import org.labkey.api.gwt.client.AuditBehaviorType;
 import org.labkey.api.pipeline.browse.PipelinePathForm;
-import org.labkey.api.query.AbstractQueryImportAction;
-import org.labkey.api.query.BatchValidationException;
 import org.labkey.api.query.FieldKey;
 import org.labkey.api.query.QueryParam;
 import org.labkey.api.query.QueryService;
 import org.labkey.api.query.QueryUpdateForm;
 import org.labkey.api.query.QueryView;
 import org.labkey.api.query.UserSchemaAction;
-import org.labkey.api.query.ValidationException;
-import org.labkey.api.reader.ColumnDescriptor;
-import org.labkey.api.reader.DataLoader;
 import org.labkey.api.security.ActionNames;
 import org.labkey.api.security.RequiresPermission;
 import org.labkey.api.security.User;
@@ -70,9 +58,9 @@ import org.labkey.api.specimen.SpecimenRequestManager;
 import org.labkey.api.specimen.SpecimenRequestManager.SpecimenRequestInput;
 import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.Vial;
-import org.labkey.api.specimen.actions.HiddenFormInputGenerator;
 import org.labkey.api.specimen.actions.IdForm;
 import org.labkey.api.specimen.actions.ParticipantCommentForm;
+import org.labkey.api.specimen.actions.SpecimensViewBean;
 import org.labkey.api.specimen.importer.RequestabilityManager;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
@@ -94,7 +82,6 @@ import org.labkey.api.specimen.security.permissions.SetSpecimenCommentsPermissio
 import org.labkey.api.specimen.settings.RequestNotificationSettings;
 import org.labkey.api.specimen.settings.SettingsManager;
 import org.labkey.api.study.CohortFilter;
-import org.labkey.api.study.Location;
 import org.labkey.api.study.SpecimenService;
 import org.labkey.api.study.SpecimenUrls;
 import org.labkey.api.study.Study;
@@ -103,19 +90,15 @@ import org.labkey.api.study.StudyUrls;
 import org.labkey.api.study.TimepointType;
 import org.labkey.api.study.Visit;
 import org.labkey.api.study.model.ParticipantDataset;
-import org.labkey.api.util.Button;
 import org.labkey.api.util.ConfigurationException;
 import org.labkey.api.util.DateUtil;
-import org.labkey.api.util.FileStream;
 import org.labkey.api.util.Link.LinkBuilder;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.util.Pair;
 import org.labkey.api.util.URLHelper;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.DataView;
-import org.labkey.api.view.DisplayElement;
 import org.labkey.api.view.GridView;
-import org.labkey.api.view.HBox;
 import org.labkey.api.view.HtmlView;
 import org.labkey.api.view.HttpView;
 import org.labkey.api.view.InsertView;
@@ -143,7 +126,6 @@ import org.springframework.validation.Errors;
 import org.springframework.validation.ObjectError;
 import org.springframework.web.servlet.ModelAndView;
 
-import javax.servlet.ServletException;
 import javax.servlet.http.HttpSession;
 import java.io.BufferedOutputStream;
 import java.io.ByteArrayOutputStream;
@@ -158,7 +140,6 @@ import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -238,27 +219,21 @@ public class SpecimenController extends BaseStudyController
         }
 
         @Override
-        public ActionURL getManageRequestURL(Container c)
+        public ActionURL getManageRequestURL(Container c, int requestId)
         {
-            return new ActionURL(ManageRequestAction.class, c);
-        }
-
-        @Override
-        public ActionURL getManageRequestStatusURL(Container c, int requestId)
-        {
-            return new ActionURL(ManageRequestStatusAction.class, c).addParameter("id", requestId);
+            return SpecimenMigrationService.get().getManageRequestURL(c, requestId, null);
         }
 
         @Override
         public ActionURL getRequestDetailsURL(Container c, int requestId)
         {
-            return new ActionURL(ManageRequestAction.class, c).addParameter("id", requestId);
+            return getManageRequestURL(c, requestId);
         }
 
         @Override
-        public ActionURL getRequestDetailsURL(Container c, String requestId)
+        public ActionURL getManageRequestStatusURL(Container c, int requestId)
         {
-            return new ActionURL(ManageRequestAction.class, c).addParameter("id", requestId);
+            return SpecimenMigrationService.get().getManageRequestStatusURL(c, requestId);
         }
 
         @Override
@@ -326,16 +301,7 @@ public class SpecimenController extends BaseStudyController
 
     private ActionURL getManageRequestURL(int requestID)
     {
-        return getManageRequestURL(requestID, null);
-    }
-
-    private ActionURL getManageRequestURL(int requestID, String returnUrl)
-    {
-        ActionURL url = new ActionURL(ManageRequestAction.class, getContainer());
-        url.addParameter(IdForm.PARAMS.id, Integer.toString(requestID));
-        if (returnUrl != null)
-            url.addParameter(ManageRequestForm.PARAMS.returnUrl, returnUrl);
-        return url;
+        return SpecimenMigrationService.get().getManageRequestURL(getContainer(), requestID, null);
     }
 
     private Set<String> getSelectionLsids()
@@ -693,594 +659,6 @@ public class SpecimenController extends BaseStudyController
             throw new UnauthorizedException();
     }
 
-    public static class ModifySpecimenRequestForm extends IdForm implements HiddenFormInputGenerator
-    {
-        public enum PARAMS
-        {
-            specimenIds,
-            returnUrl
-        }
-
-        private String _specimenIds;
-
-        public String getSpecimenIds()
-        {
-            return _specimenIds;
-        }
-
-        public void setSpecimenIds(String specimenIds)
-        {
-            _specimenIds = specimenIds;
-        }
-
-        @Override
-        public String getHiddenFormInputs(ViewContext ctx)
-        {
-            StringBuilder builder = new StringBuilder();
-            if (getId() != 0)
-                builder.append("<input type=\"hidden\" name=\"id\" value=\"").append(getId()).append("\">\n");
-            if (_specimenIds != null)
-                builder.append("<input type=\"hidden\" name=\"specimenIds\" value=\"").append(PageFlowUtil.filter(_specimenIds)).append("\">");
-            return builder.toString();
-        }
-    }
-
-    public static class ManageRequestForm extends IdForm
-    {
-        public enum PARAMS
-        {
-            newSite,
-            newActor,
-            newDescription,
-            export,
-            submissionResult,
-            returnUrl
-        }
-        private Integer _newSite;
-        private Integer _newActor;
-        private String _newDescription;
-        private String _export;
-        private Boolean _submissionResult;
-
-        public Integer getNewActor()
-        {
-            return _newActor;
-        }
-
-        public void setNewActor(Integer newActor)
-        {
-            _newActor = newActor;
-        }
-
-        public Integer getNewSite()
-        {
-            return _newSite;
-        }
-
-        public void setNewSite(Integer newSite)
-        {
-            _newSite = newSite;
-        }
-
-        public String getNewDescription()
-        {
-            return _newDescription;
-        }
-
-        public void setNewDescription(String newDescription)
-        {
-            _newDescription = newDescription;
-        }
-
-        public String getExport()
-        {
-            return _export;
-        }
-
-        public void setExport(String export)
-        {
-            _export = export;
-        }
-
-        public Boolean isSubmissionResult()
-        {
-            return _submissionResult;
-        }
-
-        public void setSubmissionResult(Boolean submissionResult)
-        {
-            _submissionResult = submissionResult;
-        }
-    }
-
-    public abstract static class SpecimensViewBean
-    {
-        protected SpecimenQueryView _specimenQueryView;
-        protected List<Vial> _vials;
-
-        public SpecimensViewBean(ViewContext context, List<Vial> vials, boolean showHistoryLinks,
-                                 boolean showRecordSelectors, boolean disableLowVialIndicators, boolean restrictRecordSelectors)
-        {
-            _vials = vials;
-            if (vials != null && vials.size() > 0)
-            {
-                _specimenQueryView = SpecimenQueryView.createView(context, vials, SpecimenQueryView.ViewType.VIALS);
-                _specimenQueryView.setShowHistoryLinks(showHistoryLinks);
-                _specimenQueryView.setShowRecordSelectors(showRecordSelectors);
-                _specimenQueryView.setDisableLowVialIndicators(disableLowVialIndicators);
-                _specimenQueryView.setRestrictRecordSelectors(restrictRecordSelectors);
-            }
-        }
-
-        public SpecimenQueryView getSpecimenQueryView()
-        {
-            return _specimenQueryView;
-        }
-
-        public List<Vial> getVials()
-        {
-            return _vials;
-        }
-    }
-
-    public class ManageRequestBean extends SpecimensViewBean
-    {
-        protected SpecimenRequest _specimenRequest;
-        protected boolean _requestManager;
-        protected boolean _requirementsComplete;
-        protected boolean _finalState;
-        private List<ActorNotificationRecipientSet> _possibleNotifications;
-        protected List<String> _missingSpecimens = null;
-        private final Boolean _submissionResult;
-        private Location[] _providingLocations;
-        private String _returnUrl;
-
-        public ManageRequestBean(ViewContext context, SpecimenRequest specimenRequest, boolean forExport, Boolean submissionResult, String returnUrl)
-        {
-            super(context, specimenRequest.getVials(), !forExport, !forExport, forExport, false);
-            _submissionResult = submissionResult;
-            _requestManager = context.getContainer().hasPermission(context.getUser(), ManageRequestsPermission.class);
-            _specimenRequest = specimenRequest;
-            _finalState = SpecimenRequestManager.get().isInFinalState(_specimenRequest);
-            _requirementsComplete = true;
-            _missingSpecimens = SpecimenRequestManager.get().getMissingSpecimens(_specimenRequest);
-            _returnUrl = returnUrl;
-            SpecimenRequestRequirement[] requirements = specimenRequest.getRequirements();
-            for (int i = 0; i < requirements.length && _requirementsComplete; i++)
-            {
-                SpecimenRequestRequirement requirement = requirements[i];
-                _requirementsComplete = requirement.isComplete();
-            }
-
-            if (_specimenQueryView != null)
-            {
-                List<DisplayElement> buttons = new ArrayList<>();
-
-                MenuButton exportMenuButton = new MenuButton("Export");
-                ActionURL exportExcelURL = context.getActionURL().clone().addParameter("export", "excel");
-                ActionURL exportTextURL = context.getActionURL().clone().addParameter("export", "tsv");
-                exportMenuButton.addMenuItem("Export all to Excel (.xls)", exportExcelURL);
-                exportMenuButton.addMenuItem("Export all to text file (.tsv)", exportTextURL);
-                buttons.add(exportMenuButton);
-                _specimenQueryView.setShowExportButtons(false);
-                _specimenQueryView.getSettings().setAllowChooseView(false);
-
-                if (SpecimenRequestManager.get().hasEditRequestPermissions(context.getUser(), specimenRequest) && !_finalState)
-                {
-                    ActionButton addButton = new ActionButton(SpecimenMigrationService.get().getOverviewURL(getContainer()), "Specimen Search");
-                    ActionButton deleteButton = new ActionButton(RemoveRequestSpecimensAction.class, "Remove Selected");
-                    _specimenQueryView.addHiddenFormField("id", "" + specimenRequest.getRowId());
-                    buttons.add(addButton);
-
-                    ActionURL importActionURL = new ActionURL(ImportVialIdsAction.class, getContainer());
-                    importActionURL.addParameter("id", specimenRequest.getRowId());
-                    Button importButton = new Button.ButtonBuilder("Upload Specimen Ids")
-                        .href(importActionURL)
-                        .submit(false)
-                        .build();
-                    buttons.add(importButton);
-                    buttons.add(deleteButton);
-                }
-                _specimenQueryView.setButtons(buttons);
-            }
-        }
-
-        public synchronized List<ActorNotificationRecipientSet> getPossibleNotifications()
-        {
-            if (_possibleNotifications == null)
-                _possibleNotifications = getUtils().getPossibleNotifications(_specimenRequest);
-            return _possibleNotifications;
-        }
-
-        public SpecimenRequest getSpecimenRequest()
-        {
-            return _specimenRequest;
-        }
-
-        public boolean hasMissingSpecimens()
-        {
-            return _missingSpecimens != null && !_missingSpecimens.isEmpty();
-        }
-
-        public List<String> getMissingSpecimens()
-        {
-            return _missingSpecimens;
-        }
-
-        public SpecimenRequestStatus getStatus()
-        {
-            return SpecimenRequestManager.get().getRequestStatus(_specimenRequest.getContainer(), _specimenRequest.getStatusId());
-        }
-
-        public Location getDestinationSite()
-        {
-            Integer destinationSiteId = _specimenRequest.getDestinationSiteId();
-            if (destinationSiteId != null)
-            {
-                return LocationManager.get().getLocation(_specimenRequest.getContainer(), destinationSiteId.intValue());
-            }
-            return null;
-        }
-
-        public boolean isRequestManager()
-        {
-            return _requestManager;
-        }
-
-        public boolean isFinalState()
-        {
-            return _finalState;
-        }
-
-        public boolean isRequirementsComplete()
-        {
-            return _requirementsComplete;
-        }
-
-        public Boolean isSuccessfulSubmission()
-        {
-            return _submissionResult != null && _submissionResult.booleanValue();
-        }
-
-        public Location[] getProvidingLocations()
-        {
-            if (_providingLocations == null)
-            {
-                Set<Integer> locationSet = new HashSet<>();
-                for (Vial vial : _vials)
-                {
-                    Integer locationId = vial.getCurrentLocation();
-                    if (locationId != null)
-                        locationSet.add(locationId);
-                }
-                _providingLocations = new Location[locationSet.size()];
-                int i = 0;
-                for (Integer locationId : locationSet)
-                    _providingLocations[i++] = LocationManager.get().getLocation(getContainer(), locationId);
-            }
-            return _providingLocations;
-        }
-
-        public String getReturnUrl()
-        {
-            return _returnUrl;
-        }
-
-        public void setReturnUrl(String returnUrl)
-        {
-            _returnUrl = returnUrl;
-        }
-    }
-
-
-    @RequiresPermission(ReadPermission.class)
-    public class ManageRequestAction extends FormViewAction<ManageRequestForm>
-    {
-        private int _requestId;
-
-        @Override
-        public void validateCommand(ManageRequestForm target, Errors errors)
-        {
-        }
-
-        @Override
-        public ModelAndView getView(ManageRequestForm form, boolean reshow, BindException errors) throws Exception
-        {
-            SpecimenRequest request = SpecimenRequestManager.get().getRequest(getContainer(), form.getId());
-            if (request == null)
-                throw new NotFoundException();
-
-            _requestId = request.getRowId();
-            ManageRequestBean bean = new ManageRequestBean(getViewContext(), request, form.getExport() != null,
-                    form.isSubmissionResult(), form.getReturnUrl());
-            if (form.getExport() != null)
-            {
-                bean.getSpecimenQueryView().getSettings().setMaxRows(Table.ALL_ROWS);   // #34998; exporting specimens in a request should include all of them
-                getUtils().writeExportData(bean.getSpecimenQueryView(), form.getExport());
-                return null;
-            }
-            else
-            {
-                GridView attachmentsGrid = getUtils().getRequestEventAttachmentsGridView(getViewContext().getRequest(), errors, _requestId);
-                SpecimenQueryView queryView = bean.getSpecimenQueryView();
-                if (null != queryView)
-                    queryView.setTitle("Associated Specimens");
-                HBox hbox = new HBox(new JspView<>("/org/labkey/study/view/specimen/manageRequest.jsp", bean), attachmentsGrid);
-                hbox.setTableWidth("");
-                return new VBox(hbox, queryView);
-            }
-        }
-
-        @Override
-        public boolean handlePost(ManageRequestForm form, BindException errors) throws Exception
-        {
-            if (!getContainer().hasPermission(getUser(), ManageRequestsPermission.class))
-                throw new UnauthorizedException("You do not have permissions to create new specimen request requirements!");
-
-            SpecimenRequest request = SpecimenRequestManager.get().getRequest(getContainer(), form.getId());
-            if (request == null)
-                throw new NotFoundException();
-
-            if (form.getNewActor() != null && form.getNewActor() > 0)
-            {
-                SpecimenRequestActor actor = SpecimenRequestRequirementProvider.get().getActor(getContainer(), form.getNewActor());
-                if (actor != null)
-                {
-                    // an actor is valid if a site has been provided for a per-site actor, or if no site
-                    // has been provided for a non-site-specific actor. The UI should enforce this already,
-                    // so this is just a backup check.
-                    boolean validActor = (actor.isPerSite() && form.getNewSite() != null && form.getNewSite() > 0) ||
-                            (!actor.isPerSite() && (form.getNewSite() == null || form.getNewSite() <= 0));
-                    if (validActor)
-                    {
-                        SpecimenRequestRequirement requirement = new SpecimenRequestRequirement();
-                        requirement.setContainer(getContainer());
-                        if (form.getNewSite() != null && form.getNewSite() > 0)
-                            requirement.setSiteId(form.getNewSite());
-                        requirement.setActorId(form.getNewActor());
-                        requirement.setDescription(form.getNewDescription());
-                        requirement.setRequestId(request.getRowId());
-                        SpecimenRequestManager.get().createRequestRequirement(getUser(), requirement, true, true);
-                    }
-                }
-            }
-            return true;
-        }
-
-        @Override
-        public ActionURL getSuccessURL(ManageRequestForm manageRequestForm)
-        {
-            return getManageRequestURL(manageRequestForm.getId(), manageRequestForm.getReturnUrl());
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-            addSpecimenRequestNavTrail(root, _requestId);
-        }
-    }
-
-    // Additional permission checking takes place before handlePost()
-    @RequiresPermission(ReadPermission.class)
-    public class RemoveRequestSpecimensAction extends FormHandlerAction<ModifySpecimenRequestForm>
-    {
-        @Override
-        public void validateCommand(ModifySpecimenRequestForm target, Errors errors)
-        {
-        }
-
-        @Override
-        public boolean handlePost(ModifySpecimenRequestForm form, BindException errors) throws Exception
-        {
-            SpecimenRequest request = SpecimenRequestManager.get().getRequest(getContainer(), form.getId());
-            requiresEditRequestPermissions(request);
-            long[] ids = toLongArray(DataRegionSelection.getSelected(getViewContext(), true));
-            List<Long> specimenIds = new ArrayList<>();
-            for (long id : ids)
-                specimenIds.add(id);
-            try
-            {
-                SpecimenRequestManager.get().deleteRequestSpecimenMappings(getUser(), request, specimenIds, true);
-            }
-            catch (RequestabilityManager.InvalidRuleException e)
-            {
-                errors.reject(ERROR_MSG, "The specimens could not be removed because a requestability rule is configured incorrectly. " +
-                            "Please report this problem to an administrator. Error details: "  + e.getMessage());
-                return false;
-            }
-            return true;
-        }
-
-        @Override
-        public ActionURL getSuccessURL(ModifySpecimenRequestForm addToSpecimenRequestForm)
-        {
-            return getManageRequestURL(addToSpecimenRequestForm.getId(), addToSpecimenRequestForm.getReturnUrl());
-        }
-    }
-
-    public static class ManageRequestStatusForm extends IdForm
-    {
-        private int _status;
-        private String _comments;
-        private String _requestDescription;
-        private String[] _notificationIdPairs;
-        private boolean  _emailInactiveUsers;
-
-        public int getStatus()
-        {
-            return _status;
-        }
-
-        public void setStatus(int status)
-        {
-            _status = status;
-        }
-
-        public String getComments()
-        {
-            return _comments;
-        }
-
-        public void setComments(String comments)
-        {
-            _comments = comments;
-        }
-
-        public String[] getNotificationIdPairs()
-        {
-            return _notificationIdPairs;
-        }
-
-        public void setNotificationIdPairs(String[] notificationIdPairs)
-        {
-            _notificationIdPairs = notificationIdPairs;
-        }
-
-        public String getRequestDescription()
-        {
-            return _requestDescription;
-        }
-
-        public void setRequestDescription(String requestDescription)
-        {
-            _requestDescription = requestDescription;
-        }
-
-        public boolean isEmailInactiveUsers()
-        {
-            return _emailInactiveUsers;
-        }
-
-        public void setEmailInactiveUsers(boolean emailInactiveUsers)
-        {
-            _emailInactiveUsers = emailInactiveUsers;
-        }
-    }
-
-    @RequiresPermission(RequestSpecimensPermission.class)
-    public class ManageRequestStatusAction extends FormViewAction<ManageRequestStatusForm>
-    {
-        private SpecimenRequest _specimenRequest;
-
-        @Override
-        public void validateCommand(ManageRequestStatusForm target, Errors errors)
-        {
-        }
-
-        @Override
-        public ModelAndView getView(ManageRequestStatusForm form, boolean reshow, BindException errors)
-        {
-            _specimenRequest = SpecimenRequestManager.get().getRequest(getContainer(), form.getId());
-            if (_specimenRequest == null)
-                throw new NotFoundException();
-
-            return new JspView<>("/org/labkey/study/view/specimen/manageRequestStatus.jsp",
-                    new ManageRequestBean(getViewContext(), _specimenRequest, false, null, null), errors);
-        }
-
-        @Override
-        public boolean handlePost(final ManageRequestStatusForm form, BindException errors) throws Exception
-        {
-            _specimenRequest = SpecimenRequestManager.get().getRequest(getContainer(), form.getId());
-            if (_specimenRequest == null)
-                throw new NotFoundException();
-
-            boolean statusChanged = form.getStatus() != _specimenRequest.getStatusId();
-            boolean detailsChanged = !nullSafeEqual(form.getRequestDescription(), _specimenRequest.getComments());
-
-            List<AttachmentFile> files = getAttachmentFileList();
-            boolean hasAttachments = !files.isEmpty();
-
-            boolean hasComments = form.getComments() != null && form.getComments().length() > 0;
-            if (statusChanged || detailsChanged || hasComments || hasAttachments)
-            {
-                RequestEventType changeType;
-                String comment = "";
-                String eventSummary;
-                if (statusChanged || detailsChanged)
-                {
-                    if (statusChanged)
-                    {
-                        SpecimenRequestStatus prevStatus = SpecimenRequestManager.get().getRequestStatus(getContainer(), _specimenRequest.getStatusId());
-                        SpecimenRequestStatus newStatus = SpecimenRequestManager.get().getRequestStatus(getContainer(), form.getStatus());
-                        comment += "Status changed from \"" + (prevStatus != null ? prevStatus.getLabel() : "N/A") + "\" to \"" +
-                                (newStatus != null ? newStatus.getLabel() : "N/A") + "\"\n";
-                    }
-                    if (detailsChanged)
-                    {
-                        String prevDetails = _specimenRequest.getComments();
-                        String newDetails = form.getRequestDescription();
-                        comment += "Description changed from \"" + (prevDetails != null ? prevDetails : "N/A") + "\" to \"" +
-                                (newDetails != null ? newDetails : "N/A") + "\"\n";
-                    }
-                    eventSummary = comment;
-                    if (hasComments)
-                        comment += form.getComments();
-                    _specimenRequest = _specimenRequest.createMutable();
-                    _specimenRequest.setStatusId(form.getStatus());
-                    _specimenRequest.setComments(form.getRequestDescription());
-                    _specimenRequest.setModified(new Date());
-                    try
-                    {
-                        SpecimenRequestManager.get().updateRequest(getUser(), _specimenRequest);
-                    }
-                    catch (RequestabilityManager.InvalidRuleException e)
-                    {
-                        errors.reject(ERROR_MSG, "The request could not be updated because a requestability rule is configured incorrectly. " +
-                                    "Please report this problem to an administrator. Error details: "  + e.getMessage());
-                        return false;
-                    }
-                    changeType = RequestEventType.REQUEST_STATUS_CHANGED;
-                }
-                else
-                {
-                    changeType = RequestEventType.COMMENT_ADDED;
-                    comment = form.getComments();
-                    eventSummary = "Comments added.";
-                }
-
-                SpecimenRequestEvent event;
-                try
-                {
-                    event = SpecimenRequestManager.get().createRequestEvent(getUser(), _specimenRequest, changeType, comment, files);
-                }
-                catch (Exception e)
-                {
-                    errors.reject(ERROR_MSG, "The request could not be updated because of an unexpected error. " +
-                            "Please report this problem to an administrator. Error details: "  + e.getMessage());
-                    return false;
-                }
-                try
-                {
-                    List<? extends NotificationRecipientSet> recipients = getUtils().getNotifications(_specimenRequest, form.getNotificationIdPairs());
-                    DefaultRequestNotification notification = new DefaultRequestNotification(_specimenRequest, recipients,
-                            eventSummary, event, form.getComments(), null, getViewContext());
-                    getUtils().sendNotification(notification, form.isEmailInactiveUsers(), errors);
-                }
-                catch (ConfigurationException | IOException e)
-                {
-                    errors.reject(ERROR_MSG, "The request was updated successfully, but the notification failed: " +  e.getMessage());
-                    return false;
-                }
-            }
-
-            return !errors.hasErrors();
-        }
-
-        @Override
-        public ActionURL getSuccessURL(ManageRequestStatusForm manageRequestForm)
-        {
-            return getManageRequestURL(_specimenRequest.getRowId());
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-            addSpecimenRequestNavTrail(root, _specimenRequest.getRowId());
-            root.addChild("Update Request");
-        }
-    }
-
     public static List<Vial> getSpecimensFromRowIds(long[] requestedSampleIds, Container container, User user)
     {
         List<Vial> requestedVials = null;
@@ -1619,7 +997,7 @@ public class SpecimenController extends BaseStudyController
         public ModelAndView getView(IdForm form, BindException errors)
         {
             _requestId = form.getId();
-            HtmlView header = new HtmlView(new LinkBuilder("View Request").href(new ActionURL(ManageRequestAction.class, getContainer()).addParameter("id", form.getId())));
+            HtmlView header = new HtmlView(new LinkBuilder("View Request").href(SpecimenMigrationService.get().getManageRequestURL(getContainer(), form.getId(), null)));
             SimpleFilter filter = new SimpleFilter(FieldKey.fromParts("RequestId"), form.getId());
             GridView historyGrid = getUtils().getRequestEventGridView(getViewContext().getRequest(), errors, filter);
             return new VBox(header, historyGrid);
@@ -1768,9 +1146,7 @@ public class SpecimenController extends BaseStudyController
         @Override
         public ActionURL getSuccessURL(IdForm idForm)
         {
-            ActionURL successURL = getManageRequestURL(idForm.getId());
-            successURL.addParameter(ManageRequestForm.PARAMS.submissionResult, Boolean.TRUE.toString());
-            return successURL;
+            return SpecimenMigrationService.get().getManageRequestURL(getContainer(), idForm.getId(), null).addParameter("submissionResult", true);
         }
     }
 
@@ -3039,181 +2415,6 @@ public class SpecimenController extends BaseStudyController
                 }
             }
             return null;
-        }
-    }
-
-    @RequiresPermission(RequestSpecimensPermission.class)
-    public class ImportVialIdsAction extends AbstractQueryImportAction<IdForm>
-    {
-        private int _requestId = -1;
-
-        @Override
-        protected void initRequest(IdForm form) throws ServletException
-        {
-            _requestId = form.getId();
-            setHasColumnHeaders(false);
-            setImportMessage("Upload a list of Global Unique Identifiers from a TXT, CSV or Excel file or paste the list directly into the text box below. " +
-                    "The list must have only one column and no header row.");
-            setNoTableInfo();
-            setHideTsvCsvCombo(true);
-        }
-
-        @Override
-        public ModelAndView getView(IdForm form, BindException errors) throws Exception
-        {
-            initRequest(form);
-            SpecimenRequest request = SpecimenRequestManager.get().getRequest(getContainer(), _requestId);
-            if (request == null)
-                throw new NotFoundException();
-
-            if (!SpecimenRequestManager.get().hasEditRequestPermissions(getUser(), request) ||
-                    SpecimenRequestManager.get().isInFinalState(request))
-            {
-                return new HtmlView("<div class=\"labkey-error\">You do not have permissions to modify this request.</div>");
-            }
-
-            return getDefaultImportView(form, errors);
-        }
-
-        @Override
-        protected void validatePermission(User user, BindException errors)
-        {
-            checkPermissions();
-            SpecimenRequest request = SpecimenRequestManager.get().getRequest(getContainer(), _requestId);
-
-            if (!SpecimenRequestManager.get().hasEditRequestPermissions(getUser(), request) ||
-                    SpecimenRequestManager.get().isInFinalState(request))
-            {
-                // No permission
-                errors.reject(SpringActionController.ERROR_MSG, "You do not have permission to modify this request.");
-            }
-        }
-
-        @Override
-        protected boolean canInsert(User user)
-        {
-            return getContainer().hasPermission(user, RequestSpecimensPermission.class);
-        }
-
-        @Override
-        protected int importData(DataLoader dl, FileStream file, String originalName, BatchValidationException errors, @Nullable AuditBehaviorType auditBehaviorType, @Nullable TransactionAuditProvider.TransactionAuditEvent auditEvent) throws IOException
-        {
-            List<String> errorList = new LinkedList<>();
-
-            ColumnDescriptor[] columns = new ColumnDescriptor[2];   // Only 1 actual column of type String
-            columns[0] = new ColumnDescriptor("Actual", String.class);
-            columns[1] = new ColumnDescriptor("Dummy", String.class);
-            dl.setColumns(columns);
-
-            List<Map<String, Object>> rows = dl.load();
-            String[] globalIds = new String[rows.size()];
-            int i = 0;
-            for (Map<String, Object> row : rows)
-            {
-                String idActual = (String)row.get("Actual");
-                if (idActual == null)
-                {
-                    errorList.add("Malformed input data.");
-                    break;
-                }
-
-                String idDummy = (String)row.get("Dummy");
-                if (idDummy != null)
-                {
-                    errorList.add("Only one id per line is allowed.");
-                    break;
-                }
-                globalIds[i] = idActual;
-                i += 1;
-            }
-
-            if (errorList.size() == 0)
-            {
-                // Get all the specimen objects. If it throws an exception then there was an error and we'll
-                // root around to figure out what to report
-                SpecimenManagerNew specimenManager = SpecimenManagerNew.get();
-                try
-                {
-                    SpecimenRequest request = SpecimenRequestManager.get().getRequest(getContainer(), _requestId);
-                    List<Vial> vials = SpecimenManagerNew.get().getVials(getContainer(), getUser(), globalIds);
-
-                    if (vials != null && vials.size() == globalIds.length)
-                    {
-                        // All the specimens exist;
-                        // Check for Availability and then add them to the request.
-                        // There still may be errors (like someone already has requested that specimen) which will be
-                        // caught by createRequestSpecimenMapping
-
-                        for (Vial s : vials)
-                        {
-                            if (!s.isAvailable()) // || s.isLockedInRequest())
-                            {
-                                errorList.add(RequestabilityManager.makeSpecimenUnavailableMessage(s, null));
-                            }
-                        }
-
-                        if (errorList.size() == 0)
-                        {
-                            ArrayList<Vial> vialList = new ArrayList<>(vials.size());
-                            vialList.addAll(vials);
-                            SpecimenRequestManager.get().createRequestSpecimenMapping(getUser(), request, vialList, true, true);
-                        }
-                    }
-                    else
-                    {
-                        errorList.add("Duplicate Ids found.");
-                    }
-                }
-                catch (RequestabilityManager.InvalidRuleException e)
-                {
-                    errorList.add("The request could not be created because a requestability rule is configured incorrectly. " +
-                            "Please report this problem to an administrator. Error details: " + e.getMessage());
-                }
-                catch (SpecimenRequestException e)
-                {
-                    // There was an error; some id had no specimen matching
-                    boolean hasSpecimenError = false;
-                    for (String id : globalIds)
-                    {
-                        Vial vial = specimenManager.getVial(getContainer(), getUser(), id);
-                        if (vial == null)
-                        {
-                            errorList.add("Specimen " + id + " not found.");
-                            hasSpecimenError = true;
-                        }
-                    }
-
-                    if (!hasSpecimenError)
-                    {   // Expected one of them to not be found, so this is unusual
-                        errorList.add("Error adding all of the specimens together.");
-                    }
-                }
-            }
-
-            if (!errorList.isEmpty())
-            {
-                for (String error : errorList)
-                    errors.addRowError(new ValidationException(error));
-            }
-
-            if (errors.hasErrors())
-                return 0;
-            return rows.size();
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-            addSpecimenRequestNavTrail(root, _requestId);
-            root.addChild("Upload Specimen Identifiers");
-        }
-
-        @Override
-        protected ActionURL getSuccessURL(IdForm form)
-        {
-            ActionURL requestDetailsURL = new ActionURL(ManageRequestAction.class, getContainer());
-            requestDetailsURL.addParameter("id", _requestId);
-            return requestDetailsURL;
         }
     }
