+++ study/src/org/labkey/study/view/specimen/labSpecimenLists.jsp
@@ -15,16 +15,16 @@
  * limitations under the License.
  */
 %>
+<%@ page import="org.labkey.api.specimen.SpecimenMigrationService"%>
 <%@ page import="org.labkey.api.specimen.Vial"%>
 <%@ page import="org.labkey.api.specimen.location.LocationImpl"%>
 <%@ page import="org.labkey.api.specimen.notifications.ActorNotificationRecipientSet"%>
 <%@ page import="org.labkey.api.view.ActionURL"%>
-<%@ page import="org.labkey.api.view.HttpView"%>
+<%@ page import="org.labkey.api.view.HttpView" %>
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.DownloadSpecimenListAction" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.EmailLabSpecimenListsAction" %>
 <%@ page import="org.labkey.study.controllers.specimen.SpecimenController.LabSpecimenListsBean" %>
-<%@ page import="org.labkey.study.controllers.specimen.SpecimenController.ManageRequestAction" %>
 <%@ taglib prefix="labkey" uri="http://www.labkey.org/taglib" %>
 <%@ page extends="org.labkey.api.jsp.JspBase" %>
 <%
@@ -192,7 +192,7 @@
     </tr>
     <tr>
         <th>&nbsp;</th>
-        <td><%= button("Send Email").submit(true) %> <%= button("Cancel").href(urlFor(ManageRequestAction.class).addParameter("id", bean.getSpecimenRequest().getRowId()))%></td>
+        <td><%= button("Send Email").submit(true) %> <%= button("Cancel").href(SpecimenMigrationService.get().getManageRequestURL(getContainer(), bean.getSpecimenRequest().getRowId(), null))%></td>
     </tr>
 </table>
 </labkey:form>
